requirements.txt
#   _____ _____ _____ _____ _____ _____ _____
#  |   __|  _  | __  |__   |   __|     |  _  |
#  |__   |     | __ -|   __|   __| | | |     |
#  |_____|__|__|_____|_____|_____|_|_|_|__|__|
#

python-dotenv
requests==2.32.3
Faker==24.14.0
